Finalizado

+++ template/app.py
@@ -16,8 +16,8 @@ class UI:
             [2] -> Remover assinatura
             [3] -> Valor total
             [4] -> Gastos últimos 12 meses
-            #Fazer o de pagamentos
-            [5] -> Sair
+            [5] -> Efetuar pagamento
+            [0] -> Sair
             ''')
 
             choice = int(input('Escolha uma opção: '))
@@ -30,6 +30,8 @@ class UI:
                 self.total_value()
             elif choice == 4:
                 self.subscription_service.gen_chart()
+            elif choice == 5:
+                self.add_payment()
             else:
                 break
 
@@ -46,15 +48,42 @@ class UI:
     def delete_subscription(self):
         subscriptions = self.subscription_service.list_all()
         print('Escolha qual assinatura deseja excluir')
+        print('0 -> Voltar')
 
         for i in subscriptions:
             print(f'{i.id} -> {i.empresa}')
 
         choice = int(input('Escolha a assinatura: '))
+
+        if choice == 0:
+            return
+
         self.subscription_service.delete(choice)
         print('Assinatura cancelada com sucesso!')
 
     def total_value(self):
         print(f'Seu valor total mensal de assinaturas é: {self.subscription_service.total_value()}')
+
+    def add_payment(self):
+        subscriptions = self.subscription_service.list_all()
+        print('Escolha qual assinatura pagar')
+        print('0 -> Voltar')
+
+        for i in subscriptions:
+            print(f'{i.id} -> {i.empresa}')
+
+        choice = int(input('Escolha a assinatura: '))
+
+        if choice == 0:
+            return
+
+        subscription = next((s for s in subscriptions if s.id == choice), None)
+
+        if subscription is None:
+            print('Assinatura não encontrada!')
+            return
+
+        self.subscription_service.pay(subscription)
+
 
 UI().start()
